+++ base/lib/qiniu_sdk_base.dart
@@ -1,4 +1,5 @@
 export 'src/auth/auth.dart';
+export 'src/client/http_client_adapter.dart';
 export 'src/error/error.dart';
 export 'src/storage/storage.dart';
 export 'src/version.dart';